pyproject.toml
[tool.poetry]
name = "user_service"
version = "0.1.0"
description = ""
authors = ["Hasan Ozgan <hasan@ozgan.net>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["opentelemetry"], version = "^0.105.0"}
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-dotenv = "^0.20.0"
pydantic = "^1.9.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.35"}
requests = "^2.27.1"
alembic = "^1.7.7"
typer = "^0.4.1"
psycopg2-binary = "^2.9.3"
object-mapper = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
flake8 = "^4.0.1"
black = "^22.3.0"
mypy = "^0.942"
autoflake = "^1.4"
pytest-cov = "^3.0.0"
coverage = "^6.3.2"
flake8-black = "^0.3.2"
isort = "^5.10.1"
sqlalchemy2-stubs = "^0.0.2-alpha.21"
autoimport = "^1.2.2"
pytest-mockito = "^0.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.poetry.scripts]
user-cli = "user_service.interfaces.cli:cli"